rename logs & reports files
