+++ src/app/manager-dashboard/manager-dashboard.module.ts
@@ -19,8 +19,8 @@ import { ReportsTableComponent } from './reports/reports-table.component';
 import { ReportsDetailComponent } from './reports/reports-detail.component';
 import { ReportsPendingComponent } from './reports/reports-pending.component';
 import { PendingTableComponent } from './reports/pending-table.component';
-import { ReportsMyPlanetComponent } from './reports/myplanet/reports.component';
-import { LogsMyPlanetComponent } from './reports/myplanet/logs.component';
+import { ReportsMyPlanetComponent } from './reports/myplanet/reports-myplanet.component';
+import { LogsMyPlanetComponent } from './reports/myplanet/logs-myplanet.component';
 import { MyPlanetToolbarComponent } from './reports/myplanet/toolbar.component';
 import { SharedComponentsModule } from '../shared/shared-components.module';
 import { ReportsDetailActivitiesComponent } from './reports/reports-detail-activities.component';
