+++ src/app/manager-dashboard/reports/myplanet/logs-myplanet.component.html
